data-minng/requirements.txt
selenium
tweepy
python-dotenv
Pillow
pandas